requirements.txt
matplotlib
pandas
tqdm
gym-factored @ git+https://github.com/tdsimao/gym-factored@master
cvxpy
